feat: разделен store на файлы
